requirements.txt
streamlit
faster-whisper
transformers==4.36.2
sentencepiece
scikit-learn
nltk
python-docx
reportlab
numpy
torch
jiwer
python-dotenv